Use dictionary lookup into of querying Core Data
Since we were fetching all entries as part of the for-loop anyway, this
shouldn't be much more memory intensive, under the assumption that
assets will not be significant compared to entries. However, we should
eventually come up with a better strategy that does not require all
entries to be in memory at once.